backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for demo project"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "azure-identity==1.19.0",
    "python-dotenv==1.0.1",
    "markitdown==0.0.2",
    "aiofiles==24.1.0",
    "aiohttp==3.11.11",
    "openai==1.78.0",
    "azure-storage-blob==12.25.0",
    "fastapi-cors==0.0.6",
    "fastapi==0.115.12",
    "uvicorn==0.34.0",
    "python-multipart==0.0.20",
    "tiktoken==0.9.0",
    "jinja2==3.1.6",
    "azure-cosmos==4.9.0",
    "azure-cognitiveservices-speech==1.43.0",
    "scipy==1.15.3"
]
